# swagger.yaml
openapi: 3.0.0
info:
  title: Donations project
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /cause/active:
    get:
      tags:
        - General
      summary: cause/active
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          example: '1'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /cause:
    post:
      tags:
        - General
      summary: cause
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: Doneaza pentru ceva
                description: Sustine ceva descriere
                target_amount: 10000
                currency: RON
                created_on: '2022-01-09 00:00:00'
                ending_on: '2022-03-09 00:00:00'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...